+++ jirust-cli/src/runners/jira_cmd_runners/version_cmd_runner.rs
@@ -1,4 +1,5 @@
 use std::collections::HashMap;
+use std::sync::Arc;
 
 use crate::args::commands::VersionArgs;
 use crate::config::config_file::{AuthData, ConfigFile};
@@ -15,6 +16,11 @@ use jira_v3_openapi::models::{
     Version, VersionRelatedWork,
 };
 use serde_json::Value;
+
+#[cfg(any(windows, unix))]
+use futures::StreamExt;
+#[cfg(any(windows, unix))]
+use futures::stream::FuturesUnordered;
 
 /// Version command runner struct
 ///
@@ -116,6 +122,135 @@ impl VersionCmdRunner {
     /// # })
     /// # }
     /// ```
+    #[cfg(any(windows, unix))]
+    pub async fn create_jira_version(
+        &self,
+        params: VersionCmdParams,
+    ) -> Result<(Version, Option<Vec<(String, String, String, String)>>), Box<dyn std::error::Error>>
+    {
+        let version_description: Option<String>;
+        let mut resolved_issues = vec![];
+        let mut transitioned_issue: Arc<Vec<(String, String, String, String)>> = Arc::new(vec![]);
+        if Option::is_some(&params.changelog_file) {
+            let changelog_extractor = ChangelogExtractor::new(params.changelog_file.unwrap());
+            version_description = Some(changelog_extractor.extract_version_changelog().unwrap_or(
+                if Option::is_some(&params.version_description) {
+                    params.version_description.unwrap()
+                } else {
+                    "No changelog found for this version".to_string()
+                },
+            ));
+            if Option::is_some(&params.transition_issues) && params.transition_issues.unwrap() {
+                resolved_issues = changelog_extractor
+                    .extract_issues_from_changelog(
+                        &version_description.clone().unwrap(),
+                        &params.project,
+                    )
+                    .unwrap_or_default();
+            }
+        } else {
+            version_description = params.version_description;
+        }
+        let release_date =
+            if Option::is_some(&params.version_released) && params.version_released.unwrap() {
+                if Option::is_some(&params.version_release_date) {
+                    params.version_release_date
+                } else {
+                    Some(Utc::now().format("%Y-%m-%d").to_string())
+                }
+            } else {
+                None
+            };
+        let version = Version {
+            project: Some(params.project),
+            name: Some(
+                params
+                    .version_name
+                    .expect("VersionName is mandatory on creation!"),
+            ),
+            description: version_description,
+            start_date: params.version_start_date,
+            release_date,
+            archived: params.version_archived,
+            released: params.version_released,
+            ..Default::default()
+        };
+        let version = create_version(&self.cfg, version).await?;
+        if !resolved_issues.is_empty() {
+            let user_data = if Option::is_some(&params.transition_assignee) {
+                Some(User {
+                    account_id: Some(params.transition_assignee.expect("Assignee is required")),
+                    account_type: Some(AccountType::Atlassian),
+                    ..Default::default()
+                })
+            } else {
+                None
+            };
+            let mut handles = FuturesUnordered::new();
+            for issue in resolved_issues {
+                handles.push(self.manage_version_related_issues(issue, &user_data, &version));
+            }
+            while let Some(result) = handles.next().await {
+                match result {
+                    Ok((issue, transition_result, assign_result, fixversion_result)) => {
+                        Arc::make_mut(&mut transitioned_issue).push((
+                            issue,
+                            transition_result,
+                            assign_result,
+                            fixversion_result,
+                        ));
+                    }
+                    Err(err) => {
+                        eprintln!("Error managing version related issues: {err:?}");
+                    }
+                }
+            }
+        }
+        let transitioned_issue_owned: Vec<(String, String, String, String)> =
+            (*transitioned_issue).clone();
+        Ok((
+            version,
+            if !transitioned_issue.is_empty() {
+                Some(transitioned_issue_owned)
+            } else {
+                None
+            },
+        ))
+    }
+
+    /// This method creates a new Jira version with the given parameters
+    /// and returns the created version
+    ///
+    /// # Arguments
+    ///
+    /// * `params` - A VersionCmdParams struct
+    ///
+    /// # Returns
+    ///
+    /// * A Result containing a Version struct or a Box<dyn std::error::Error>
+    ///
+    /// # Examples
+    ///
+    /// ```no_run
+    /// use jira_v3_openapi::models::Version;
+    /// use jirust_cli::runners::jira_cmd_runners::version_cmd_runner::VersionCmdParams;
+    /// use jirust_cli::config::config_file::ConfigFile;
+    /// use jirust_cli::runners::jira_cmd_runners::version_cmd_runner::VersionCmdRunner;
+    /// use toml::Table;
+    ///
+    /// # fn main() -> Result<(), Box<dyn std::error::Error>> {
+    /// # tokio_test::block_on(async {
+    /// let cfg_file = ConfigFile::new("dXNlcm5hbWU6YXBpX2tleQ==".to_string(), "jira_url".to_string(), "standard_resolution".to_string(), "standard_resolution_comment".to_string(), Table::new());
+    ///
+    /// let version_cmd_runner = VersionCmdRunner::new(&cfg_file);
+    /// let params = VersionCmdParams::new();
+    ///
+    /// let version = version_cmd_runner.create_jira_version(params).await?;
+    /// # Ok(())
+    /// # })
+    /// # }
+    /// ```
+    #[cfg(target_family = "wasm")]
     pub async fn create_jira_version(
         &self,
         params: VersionCmdParams,
@@ -582,6 +717,168 @@ impl VersionCmdRunner {
             params.version_id.expect("VersionID is mandatory!").as_str(),
         )
         .await
+    }
+
+    /// Manage version related issues helper function
+    /// Use FuturesUnordered to manage multiple operation concurrently
+    ///
+    /// # Arguments
+    /// * `issue` - The issue to manage
+    /// * `user_data` - The user data
+    /// * `version` - The version
+    ///
+    /// # Returns
+    /// A Result containing a tuple with the issue key, status, resolution, and fix version or an error
+    ///
+    /// # Example
+    /// ```ignore
+    /// use jira_v3_openapi::models::{ User, Version };
+    /// use jirust_cli::config::config_file::ConfigFile;
+    /// use jirust_cli::runners::jira_cmd_runners::version_cmd_runner::VersionCmdRunner;
+    /// use toml::Table;
+    ///
+    /// let cfg_file = ConfigFile::new("dXNlcm5hbWU6YXBpX2tleQ==".to_string(), "jira_url".to_string(), "standard_resolution".to_string(), "standard_resolution_comment".to_string(), Table::new());
+    /// let version_cmd_runner = VersionCmdRunner::new(&cfg_file);
+    /// let issue_key = "ABC-123";
+    /// let user_data = Some(User::default());
+    /// let version = Version::default();
+    /// let result = self.manage_version_related_issues(issue_key, &user_data, &version).await;
+    /// ```
+    ///
+    #[cfg(any(windows, unix))]
+    async fn manage_version_related_issues(
+        &self,
+        issue: String,
+        user_data: &Option<User>,
+        version: &Version,
+    ) -> Result<(String, String, String, String), Box<dyn std::error::Error>> {
+        let all_transitions: Vec<IssueTransition> = match get_transitions(
+            &self.cfg,
+            issue.clone().as_str(),
+            None,
+            None,
+            None,
+            Some(false),
+            None,
+        )
+        .await
+        {
+            Ok(transitions) => transitions.transitions.unwrap_or_default(),
+            Err(_) => {
+                return Ok((
+                    issue,
+                    "KO".to_string(),
+                    "KO".to_string(),
+                    "NO fixVersion set".to_string(),
+                ));
+            }
+        };
+        let transition_names: Vec<String> = self
+            .resolution_transition_name
+            .clone()
+            .expect("Transition name is required and must be set in the config file");
+        let resolve_transitions: Vec<IssueTransition> = all_transitions
+            .into_iter()
+            .filter(|t| transition_names.contains(&t.name.clone().unwrap_or("".to_string())))
+            .collect();
+        let transition_ids = resolve_transitions
+            .into_iter()
+            .map(|t| t.id.clone().unwrap_or("".to_string()))
+            .collect::<Vec<String>>();
+        let transitions = transition_ids
+            .into_iter()
+            .map(|id| {
+                Some(IssueTransition {
+                    id: Some(id),
+                    ..Default::default()
+                })
+            })
+            .collect::<Vec<Option<IssueTransition>>>();
+        let mut update_fields_hashmap: HashMap<String, Vec<FieldUpdateOperation>> = HashMap::new();
+        let mut transition_fields_hashmap: HashMap<String, Vec<FieldUpdateOperation>> =
+            HashMap::new();
+        let mut version_update_op = FieldUpdateOperation::new();
+        let mut version_resolution_update_field = HashMap::new();
+        let mut version_resolution_comment_op = FieldUpdateOperation::new();
+        let version_json: Value =
+            serde_json::from_str(serde_json::to_string(&version).unwrap().as_str())
+                .unwrap_or(Value::Null);
+        let resolution_value = self.resolution_value.clone();
+        let comment_value = self.resolution_comment.clone();
+        version_update_op.add = Some(Some(version_json));
+        version_resolution_update_field.insert("resolution".to_string(), resolution_value);
+        version_resolution_comment_op.add = Some(Some(comment_value));
+        update_fields_hashmap.insert("fixVersions".to_string(), vec![version_update_op]);
+        transition_fields_hashmap
+            .insert("comment".to_string(), vec![version_resolution_comment_op]);
+        let issue_update_data = IssueUpdateDetails {
+            fields: None,
+            history_metadata: None,
+            properties: None,
+            transition: None,
+            update: Some(update_fields_hashmap),
+        };
+        let mut transition_result: String = "KO".to_string();
+        if !Vec::is_empty(&transitions) {
+            for transition in transitions {
+                let issue_transition_data = IssueUpdateDetails {
+                    fields: Some(version_resolution_update_field.clone()),
+                    history_metadata: None,
+                    properties: None,
+                    transition: Some(transition.clone().unwrap()),
+                    update: Some(transition_fields_hashmap.clone()),
+                };
+                match do_transition(&self.cfg, issue.clone().as_str(), issue_transition_data).await
+                {
+                    Ok(_) => {
+                        transition_result = "OK".to_string();
+                        break;
+                    }
+                    Err(Error::Serde(e)) => {
+                        if e.is_eof() {
+                            transition_result = "OK".to_string();
+                            break;
+                        } else {
+                            transition_result = "KO".to_string()
+                        }
+                    }
+                    Err(_) => transition_result = "KO".to_string(),
+                }
+            }
+        }
+        let assign_result: String = match assign_issue(
+            &self.cfg,
+            issue.clone().as_str(),
+            user_data.clone().unwrap(),
+        )
+        .await
+        {
+            Ok(_) => "OK".to_string(),
+            Err(Error::Serde(e)) => {
+                if e.is_eof() {
+                    "OK".to_string()
+                } else {
+                    "KO".to_string()
+                }
+            }
+            Err(_) => "KO".to_string(),
+        };
+        let fixversion_result: String = match edit_issue(
+            &self.cfg,
+            issue.clone().as_str(),
+            issue_update_data,
+            Some(true),
+            None,
+            None,
+            Some(true),
+            None,
+        )
+        .await
+        {
+            Ok(_) => version.clone().name.unwrap_or("".to_string()),
+            Err(_) => "NO fixVersion set".to_string(),
+        };
+        Ok((issue, transition_result, assign_result, fixversion_result))
     }
 }